Move all proxy logic to proxy.go
